refactor(classNameUpdater utils): move to a shared helpers folder
