+++ helpers/utils.ts
@@ -1,4 +1,4 @@
-import { readdir }  from "fs/promises";
+import { readdir } from "fs/promises";
 
 export async function isDir(potentialDirPath: string) {
   return readdir(potentialDirPath).then(
