+++ packages/class-name-updater/src/cli.ts
@@ -3,7 +3,7 @@ import { join } from "path";
 import { Command } from "commander";
 const program = new Command();
 
-import { isDir } from "./utils";
+import { isDir } from "helpers/utils";
 import { classNameUpdate } from "./classNameUpdate";
 
 program
